style: hook 파일 분리
