+++ hooks/peerConnection.ts
@@ -1,0 +1,6 @@
+import { useMeeting } from '@/hooks/meeting';
+
+export const usePeerConnections = () => {
+	const { peerConnectionsRef } = useMeeting();
+	return { peerConnectionsRef: peerConnectionsRef };
+};
